Some additional code cleanup work for EventSource.
